Input the existing project
This is the unity project of the WoodWork External that was a copy of from a Tutorial project that I Gary Au was working on.

Right now there the tools and wood is grapable.
The gluetube is able to create glue splats on the surface of the wood it muzzle was aming.
The slicing function is off due to the upgrade of the code, but will be online shortly.

Right now the project is aiming to make parts of woods to attach to each other on the side the glue splat was facing at and make a complete product object of the pre sliced wood parts.

The obstacle right now is the XR grap component and rigidbody that messing around or else we can get an alfa stage box product finnished with only glue.

Author Gary Au